move remote data to src/utils/api.ts
