+++ template/src/api/authSlice.test.ts
@@ -2,8 +2,8 @@ import { REHYDRATE } from 'redux-persist'
 import { expectSaga } from 'redux-saga-test-plan'
 import * as matchers from 'redux-saga-test-plan/matchers'
 import { throwError } from 'redux-saga-test-plan/providers'
-import { Failure, Loading, RemoteDataStates, Success } from '@api/RemoteData'
 import { resetStore } from '@redux/rootActions'
+import { Failure, Loading, RemoteDataStates, Success } from '@utils/api'
 import { logIn } from './auth'
 import {
   authSlice,
