+++ template/src/screens/DemoScreen.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import { RemoteDataStates } from '@api/RemoteData'
 import { TestIDs } from '@config/testIDs'
 import Routes from '@navigation/routes'
+import { RemoteDataStates } from '@utils/api'
 import { createNavigationProps, fireEvent, render } from '@utils/testing'
 import DemoScreen from './DemoScreen'
