+++ template/src/screens/LoginScreen.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { ActivityIndicator, Button, StyleSheet } from 'react-native'
-import { isLoading } from '@api/RemoteData'
 import { logInAsync, selectAuthTokens } from '@api/authSlice'
 import DemoTextInput from '@components/inputs/DemoTextInput'
 import MainScreenLayout from '@components/layouts/MainScreenLayout'
@@ -10,6 +9,7 @@ import useAppDispatch from '@hooks/useAppDispatch'
 import useAppSelector from '@hooks/useAppSelector'
 import { RootStackScreenProps } from '@navigation/navigators/RootStackNavigator'
 import Routes from '@navigation/routes'
+import { isLoading } from '@utils/api'
 
 const username = 'FAKE_USERNAME'
 const password = 'FAKE_PASSWORD'
